perf(images): implement multiple optimizations to limit expensive computations
The imageBoxDimensions are now computed when required. In addition,
requirements are also recomputed when necessary, preventing expensive
operations from happening too often. Also, because we need to flatten
the style prop to infer requirements, this flatten value is now cached
and re-evaluated when appropriate. Finally, styles have been moved to a
stylesheet when that is possible, to avoid commiting updates to the native
side. Note that these performance optimizations are made possible by the
high coverage rate of the HTMLImage component.